#89 fix: 임시로 작성자 이름으로 비교하여 더보기 버튼 노출 여부 결정함
