+++ src/services/party.ts
@@ -1,7 +1,7 @@
 import { PageData, Party, PartyDetail } from '@/pages/types/api'
 import api from '../utils/api'
 import { stringify } from '@/utils/query'
-import { useSuspenseInfiniteQuery, useSuspenseQuery } from '@tanstack/react-query'
+import { useQuery, useSuspenseInfiniteQuery, useSuspenseQuery } from '@tanstack/react-query'
 import { queryClient } from '@/utils/tanstack'
 import {
   ClimbingTypeEn,
@@ -284,7 +284,18 @@ export const get_party_$partyId_detail = async (partyId: number) => {
 
 export const PARTY_DETAIL_KEY = ['party', 'detail']
 
-export const usePartyDetail = (partyId: number) => {
+export const usePartyDetail = (partyId?: number) => {
+  return useQuery({
+    queryKey: [...PARTY_DETAIL_KEY, partyId],
+    queryFn: () => get_party_$partyId_detail(partyId!),
+    retryOnMount: false,
+    refetchOnWindowFocus: false,
+    select: (data) => new PartyDetailAdapter(data).adapt(),
+    enabled: !!partyId,
+  })
+}
+
+export const usePartyDetailSuspense = (partyId: number) => {
   return useSuspenseQuery({
     queryKey: [...PARTY_DETAIL_KEY, partyId],
     queryFn: () => get_party_$partyId_detail(partyId),
@@ -293,6 +304,7 @@ export const usePartyDetail = (partyId: number) => {
     select: (data) => new PartyDetailAdapter(data).adapt(),
   })
 }
+
 
 export type PartyDetailType = ReturnType<PartyDetailAdapter['adapt']>
